修正product_service無法讀取dotenv問題
